requirements.txt
mcp
fastapi
uvicorn
httpx
python-dotenv